+++ src/Resources/contao/languages/en/tl_content.php
@@ -22,5 +22,22 @@ $GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderTpl']              = [
     'Template for recursive download folder',
     'Please select the template for the recursive download folder.',
 ];
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderMode']             = [
+    'Display mode',
+    'Please choose the mode how the recursive module should work.',
+];
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderVisibleRoot']      = [
+    'Render root folder',
+    'Force rendering of root folder (Always rendered if more than one folder is selected).',
+];
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderAllowAll']         = [
+    'Ignore setting allowed downloads',
+    'List all files no matter if they are defined in the ignore download setting',
+];
 
 $GLOBALS['TL_LANG']['tl_content']['recursive-download-folder_legend'] = 'Recursive download folder settings';
+
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderModes']['breadcrumb'][0] = 'Folder navigation with breadcrumb';
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderModes']['breadcrumb'][1] = 'Folder navigation with breadcrumb.';
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderModes']['toggleable'][0] = 'Toggleable folder tree';
+$GLOBALS['TL_LANG']['tl_content']['recursiveDownloadFolderModes']['toggleable'][1] = 'All folders are toggleable without a page reload.';